Refactor prose and structure further
Pull prose that establishes convention or style into top-level
readme for language for more natural reading for authors and
readers. Fix dashes to use &mdash; for more correct typography,
and rename directory to not use number to enforce order, as order
should be enforced via the TOC in each language "book," rather
than hacked-in lexical ordering.